requirements.txt
PyQt5
psutil
requests
scikit-learn
pandas
numpy
joblib
yara-python
scapy
pywin32
cryptography
reportlab